+++ src/components/main/Banner.tsx
