+++ src/components/main/CommunitySection.tsx
