+++ src/components/main/NewsSection.tsx
