+++ src/components/main/QuizSection.tsx
